+++ Dayana/Shared/Persistence/Models/Blog/Queries/PostCategoryQueries/GetPostCategoryByFilterQuery.cs
@@ -5,13 +5,3 @@ using MediatR;
 
 namespace Dayana.Shared.Persistence.Models.Blog.Queries.BlogPosts.PostCategoryQueries;
 
-public class GetWeblogPostCategoryByFilterQuery : IRequestInfo, IRequest<OperationResult>
-{
-    public GetWeblogPostCategoryByFilterQuery(RequestInfo requestInfo)
-    {
-        RequestInfo = requestInfo;
-    }
-
-    public DefaultPaginationFilter Filter { get; set; }
-    public RequestInfo RequestInfo { get; private set; }
-}
