+++ Dayana/Shared/Persistence/Models/Blog/Queries/PostQueries/GetPostByFilterQuery.cs
@@ -5,13 +5,3 @@ using MediatR;
 
 namespace Dayana.Shared.Persistence.Models.Blog.Queries.BlogPosts.PostQueries;
 
-public record GetPostByFilterQuery : IRequestInfo, IRequest<OperationResult>
-{
-    public GetPostByFilterQuery(RequestInfo requestInfo)
-    {
-        RequestInfo = requestInfo;
-    }
-
-    public DefaultPaginationFilter Filter { get; set; }
-    public RequestInfo RequestInfo { get; private set; }
-}
