+++ Dayana/Shared/Persistence/Models/Blog/Queries/PostQueries/GetPostByIdQuery.cs
@@ -4,13 +4,3 @@ using MediatR;
 
 namespace Dayana.Shared.Persistence.Models.Blog.Queries.BlogPosts.PostQueries;
 
-public class GetPostByIdQuery : IRequestInfo, IRequest<OperationResult>
-{
-    public GetPostByIdQuery(RequestInfo requestInfo)
-    {
-        RequestInfo = requestInfo;
-    }
-
-    public int WeblogPostId { get; set; }
-    public RequestInfo RequestInfo { get; private set; }
-}
